% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSPoTApi.r
\name{ExpressionSPoTApi_add_atomic_expression}
\alias{ExpressionSPoTApi_add_atomic_expression}
\title{Create a single new object from a given data file with optional supplied metadata.}
\usage{
ExpressionSPoTApi_add_atomic_expression(source, ...)
}
\description{
Create a single new object from a given data file with optional supplied metadata.
}
