% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSPoTApi.r
\name{ExpressionSPoTApi_add_expression}
\alias{ExpressionSPoTApi_add_expression}
\title{Create multiple new objects from a multi-row data file with optional supplied metadata}
\usage{
ExpressionSPoTApi_add_expression(source, ...)
}
\arguments{
\item{source}{}
}
\description{
Create multiple new objects from a multi-row data file with optional supplied metadata
}
