% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSPoTApi.r
\name{ExpressionSPoTApi_delete_atomic}
\alias{ExpressionSPoTApi_delete_atomic}
\title{Delete the object}
\usage{
ExpressionSPoTApi_delete_atomic(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Delete the object
}
