% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSPoTApi.r
\name{ExpressionSPoTApi_get_expression_data}
\alias{ExpressionSPoTApi_get_expression_data}
\title{Retrieve multiple expression data and metadata objects}
\usage{
ExpressionSPoTApi_get_expression_data(
  filter,
  query,
  search_specific_terms,
  run_filter,
  min_expression_level,
  feature_list,
  use_versions,
  returned_metadata_fields,
  page_limit,
  cursor,
  ...
)
}
\arguments{
\item{filter}{Filter by expression metadata (key-value metadata pair(s)). E.g. \code{"Normalization Method"=TPM}.}

\item{query}{Search for expression objects via a full text query over all expression metadata. E.g. \code{TPM}.}

\item{search_specific_terms}{If the full-text query term is present in an ODM dictionary, enabling this parameter will modify the query to include child terms of the full-text query.  For example, the search query "Body fluid" can be expanded to include the term "Blood" (a child term of  "Body fluid") so files containing either "Body fluid" or  "Blood"  in their metadata will be returned in the search results.  The parent-child relationship is defined by the key "broaders" in the dictionary.  If the full query term is not present in a dictionary then this parameter has no effect.}

\item{run_filter}{Genestack accession that corresponds to the gct column used to link expression data from the same run to a sample.}

\item{min_expression_level}{Minimum threshold (inclusive) for returned expression values.}

\item{feature_list}{List of features to return. These features must match exactly the probe IDs in the GCT file. Example: \code{featureList=1552256_a_at&featureList=1552269_at}}

\item{use_versions}{Specify which versions of omics data files are used in the query. By default the active version is used. See Versioning above. Syntax:  \\* or \code{v<version number>} or \code{<CHAIN_ID>}:\code{v<version number>} or \code{<CHAIN_ID>}:\code{<accession1,accession2,..>}}

\item{returned_metadata_fields}{The parameter defines amount of metadata attributes to return:  1. \code{template} - return metadata attributes according to applied template, if object doesn’t have applied template default template will be used. This is the default for User endpoints. 2. \code{all} - return all metadata attributes with values and null attributes, if they are present in the applied template. This is the default for Curator endpoints.}

\item{page_limit}{Maximum number of results to return per page (see Paging above). This value must be between 0 and 2000 (inclusive). The default is 2000.}

\item{cursor}{The page tag to resume results from (see paging above).}
}
\description{
Retrieve multiple expression data and metadata objects
}
