% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSPoTApi.r
\name{ExpressionSPoTApi_get_expression_versions}
\alias{ExpressionSPoTApi_get_expression_versions}
\title{Retrieve a list of object versions by ID}
\usage{
ExpressionSPoTApi_get_expression_versions(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Retrieve a list of object versions by ID
}
