% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSPoTApi.r
\name{ExpressionSPoTApi_update_expression_run}
\alias{ExpressionSPoTApi_update_expression_run}
\title{Update object metadata}
\usage{
ExpressionSPoTApi_update_expression_run(id, body, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{body}{Metadata in the form of \code{{key: value, key2: value2, ...}}}
}
\description{
Update object metadata
}
