% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSPoTApi.r
\name{ExpressionSPoTApi_update_group}
\alias{ExpressionSPoTApi_update_group}
\title{Set curation status of a group by ID (accession)}
\usage{
ExpressionSPoTApi_update_group(id, body, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Set curation status of a group by ID (accession)
}
